+++ frontend/app/src/components/settings/data-security/data-management/PurgeImagesCache.vue
@@ -26,7 +26,7 @@ const ensNamesList: ComputedRef<string[]> = computed(
 );
 
 const { clearIconCache } = useAssetIconApi();
-const { setLastRefreshedAssetIcon } = useAssetIcon();
+const { setLastRefreshedAssetIcon } = useAssetIconStore();
 const { clearEnsAvatarCache } = useAddressesNamesApi();
 const { setLastRefreshedAvatar } = useAddressesNamesStore();