+++ rotkehlchen/tasks/utils.py
@@ -32,6 +32,7 @@ def should_run_periodic_task(
             DBCacheStatic.LAST_AAVE_V3_ASSETS_UPDATE,
             DBCacheStatic.LAST_DELETE_PAST_CALENDAR_EVENTS,
             DBCacheStatic.LAST_CREATE_REMINDER_CHECK_TS,
+            DBCacheStatic.LAST_GRAPH_DELEGATIONS_CHECK_TS,
         ],
         refresh_period: int,
 ) -> bool: