+++ docs/changelog.rst
@@ -2,10 +2,16 @@
 Changelog
 =========
 
-* :feature:`6115` Now free users can filter history events too. 
+* :feature:`6115` Now free users can filter history events too.
 * :feature:`7570` Users can choose whether to automatically force-push when a time discrepancy warning occurs during automatic database sync.
 * :feature:`-` Degen airdrop 2 season 3 claim event should now be properly decoded in the history events view.
 * :feature:`-` The eigenlayer airdrop claim event should now be properly decoded in the history events view.
+* :feature:`7358` Added support for detection of staked GRT tokens on Arbitrum.
+* :bug:`-` Fix the issue where the MATIC amount always shows as zero in the Polygon balance table.
+* :bug:`7915` Show OKX balances locked in active trades.
+* :feature:`-` rotki now optionally can use the decentralized TheGraph service. You can get a free api key at https://thegraph.com/studio/apikeys/
+* :feature:`7816` rotki should now support the new coinbase api key format from the coinbase develop portal.
+* :bug:`-` Compound v2 transactions containing also flash loans of same asset type will now be properly decoded.
 
 * :release:`1.33.0 <2024-05-08>`
 * :feature:`7798` rotki now accurately decodes transactions on the Kyber swap aggregator across all supported chains.